+++ packages/core/types/src/model.ts
@@ -1,19 +1,21 @@
-import { ContentKeyName } from "./compTag"
+/* eslint-disable @typescript-eslint/no-unused-vars */
+import { ContentKeyName, ContentProp } from "./compTag"
 
 type RemoveDLightLifecycle<T> = Omit<
   T,
   "willMount" | "didMount" | "didUpdate" | "willUnmount"
 >
 
-type ObjectLengthZero<T> = keyof T extends never ? true : false
-
 export type Modeling<Model, Props = object> = {
-  modeling: ContentKeyName<Props> extends never
-    ? ObjectLengthZero<Props> extends true
-      ? () => RemoveDLightLifecycle<Model>
-      : (props: Props) => RemoveDLightLifecycle<Model>
-    : (
-        props: Omit<Props, ContentKeyName<Props>>,
-        content: ContentKeyName<Props>
-      ) => RemoveDLightLifecycle<Model>
+  model: RemoveDLightLifecycle<Model>
+  props: keyof Props extends never
+    ? never
+    : ContentKeyName<Props> extends undefined
+      ? Props
+      : Omit<Props, ContentKeyName<Props>>
+  content: ContentKeyName<Props> extends undefined
+    ? never
+    : Props[ContentKeyName<Props>] extends ContentProp<infer U>
+      ? U
+      : never
 }